% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.data.frame.evalGrid}
\alias{as.data.frame.evalGrid}
\title{Converts an 'evalGrid' object into a data.frame}
\usage{
\method{as.data.frame}{evalGrid}(x, ..., value.fun = identity,
  post.proc = NULL, progress = FALSE)
}
\arguments{
\item{x}{an object returned by
\code{\link{evalGrids}}}

\item{\dots}{serves only to force the user to use the
fullnames of following two parameters.}

\item{value.fun}{A functions that converts the result
object into a data.frame}

\item{post.proc}{Functions to summarize the results over
the replications, e.g. mean, sd.}

\item{progress}{if \code{TRUE} a progress bar is shown in the console.}
}
\value{
A \code{data.frame} with the parameter constellations
 for the data generation and evaluation and the results
 (probably summarized) in \code{data.frame} manner.
}
\description{
Converts the results contained in the
 object returned by \code{\link{evalGrids}}
 into a \code{data.frame}. If the results can not
 be coerced into a data.frame automatically, the
 user can provide a function. Furthermore, functions
 to summarize the results over the replications
 the user may provide a function.
}
\details{
If necessary, more details than the description above
}
\examples{
genRegData <- function(){
 data.frame(
     x = 1:10,
     y = rnorm(10, mean=1:10))
}
# the first parameter is ALWAYS used
# for the generated data

LM <- function(data, formula){
 lm(formula=as.formula(formula), data=data)
}
eg <- evalGrids(
 expandGrid(fun="genRegData"),
 expandGrid(proc="LM", formula=c("y ~ x", "y ~ x + I(x^2)")),
 replications=5)

lm2df = function(lm.object) {
 ret = coef(summary.lm(lm.object))[, 1:2]
 data.frame(covariable = rownames(ret), ret, check.names=FALSE)
}
as.data.frame(eg, value.fun=lm2df, progress=TRUE)
as.data.frame(eg, value.fun=lm2df, post.proc=c(mean, sd), progress=TRUE)
}
\author{
Marsel Scheer
}
\seealso{
\code{\link{evalGrids}}
}

