% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_tibbles.R
\name{eval_tibbles}
\alias{eval_tibbles}
\title{Workhorse for simulation studies}
\usage{
eval_tibbles(dataGrid, procGrid = expandGrid(proc = "length"),
  replications = 1, discardGeneratedData = FALSE, post_analyze = identity,
  summary_fun = NULL, ncpus = 1L, cluster = NULL,
  clusterSeed = rep(12345, 6), clusterLibraries = NULL,
  clusterGlobalObjects = NULL, envir = globalenv())
}
\arguments{
\item{dataGrid}{a \code{data.frame} where the first column
is a character vector with function names. The other
columns contain parameters for the functions specified
in the first column. Parameters with NA are ignored.}

\item{procGrid}{similar as \code{dataGrid} the first
column must contain function names. The other columns 
contain parameters for the functions specified in the
first column. The data generated according to 
\code{dataGrid} will always be passed to the first
unspecified argument of the functions sepcified in the first 
column of \code{procGrid}.}

\item{replications}{number of replications for the simulation}

\item{discardGeneratedData}{if \code{TRUE} the generated 
data is deleted after all function constellations in 
\code{procGrid} have been applied. Otherwise, ALL
generated data sets will be part of the returned object.}

\item{summary_fun}{named list of univariate function to summarize the results (numeric or logical) over
the replications, e.g. list(mean = mean, sd = sd).}

\item{ncpus}{a cluster of \code{ncpus} workers (R-processes)
is created on the local machine to conduct the
simulation. If \code{ncpus}
equals one no cluster is created and the simulation
is conducted by the current R-process.}

\item{cluster}{a cluster generated by the \code{parallel}
package that will be used to conduct the simulation.
If \code{cluster} is specified, then \code{ncpus} will
be ignored.}

\item{clusterSeed}{if the simulation is done in parallel
manner, then the combined multiple-recursive generator from L'Ecuyer (1999) 
is used to generate random numbers. Thus \code{clusterSeed} must be a (signed) integer 
vector of length 6. The 6 elements of the seed are internally regarded as 
32-bit unsigned integers. Neither the first three nor the last three 
should be all zero, and they are limited to less than 4294967087 and 
4294944443 respectively.}

\item{clusterLibraries}{a character vector specifying
the packages that should be loaded by the workers.}

\item{clusterGlobalObjects}{a character vector specifying
the names of R objects in the global environment that should
be exported to the global environment of every worker.}

\item{envir}{must be provided if the functions specified
in \code{dataGrid} or \code{procGrid} are not part
of the global environment.}
}
\value{
The returned object list of the class
 \code{eval_tibbles}, where the element \code{simulations} contain
 the results of the simulation.
}
\description{
Generates data according to all provided 
 constellations in \code{data_tibble} and applies
 all provided constellations in \code{proc_tibble}
 to them.
}
\note{
If \code{cluster} is provided by the user the
 function \code{eval_tibbles} will NOT stop the cluster.
 This has to be done by the user. Conducting parallel
 simulations by specifing \code{ncpus} will interally
 create a cluster and stop it after the simulation
 is done.
}
\seealso{
\code{\link{as.data.frame.evalGrid}}
}
\author{
Marsel Scheer
}
