% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_tibble.R
\name{expand_tibble}
\alias{expand_tibble}
\title{Creates a \code{tibble} from All Combinations}
\usage{
expand_tibble(...)
}
\arguments{
\item{\dots}{vectors, factors or a list containing these.}
}
\value{
See \code{\link{expand.grid}} but instead of a \code{\link{data.frame}}
 a \code{\link{tibble}} is returned.
}
\description{
Actually a wrapper for \code{\link{expand.grid}}, but
 character vectors will stay as characters.
}
\examples{

expand_tibble(fun="rnorm", mean=1:4, sd=2:5)

}
\seealso{
\code{\link{expand.grid}}
}
\author{
Marsel Scheer
}
