% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandGrid.R
\name{expandGrid}
\alias{expandGrid}
\title{Creates a \code{data.frame} from All Combinations}
\usage{
expandGrid(...)
}
\arguments{
\item{\dots}{vectors, factors or a list containing these.}
}
\value{
See \code{\link{expand.grid}}
}
\description{
Actually a wrapper for \code{\link{expand.grid}}, but
 character vectors will stay as characters.
}
\examples{

expandGrid(fun="rnorm", mean=1:4, sd=2:5)

}
\seealso{
\code{\link{expand.grid}}
}
\author{
Marsel Scheer
}
