% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hardCode}
\alias{hardCode}
\title{Generates source code for the simulation}
\usage{
hardCode(dataGrid, procGrid, replications = 1, discardGeneratedData = TRUE)
}
\arguments{
\item{dataGrid}{see \code{\link{evalGrids}}}

\item{procGrid}{see \code{\link{evalGrids}}}

\item{replications}{see \code{\link{evalGrids}}}

\item{discardGeneratedData}{see \code{\link{evalGrids}}}
}
\value{
see \code{\link{evalGrids}}
}
\description{
Generates hard coded source code from the information
 provided by \code{dataGrid} and \code{procGrid}.
}
\details{
If necessary, more details than the description above
}
\examples{
\dontrun{hardCode(expandGrid(fun="runif", n=1:3),expandGrid(proc=c("rng", "median")),rep=3)}
}
\author{
Marsel Scheer
}
\seealso{
\code{\link{evalGrids}}
}

