% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hardCodeOneLine}
\alias{hardCodeOneLine}
\title{Helperfunction for \code{\link{hardCode}}}
\usage{
hardCodeOneLine(dataGrid, var, dataArgument = FALSE, silent = TRUE)
}
\arguments{
\item{dataGrid}{Describe \code{dataGrid} here}

\item{var}{Describe \code{var} here}

\item{dataArgument}{Describe \code{dataArgument} here}

\item{silent}{Describe \code{silent} here}
}
\value{
~Describe the value returned  If it is a LIST, use
\item{comp1 }{Description of 'comp1'}  \item{comp2 }{Description of
'comp2'}  ...
}
\description{
A concise (1-5 lines) description of what the function does.
}
\details{
If necessary, more details than the description above
}
\note{
further notes
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--  or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (dataGrid, var, dataArgument = FALSE, silent = TRUE)
{
   s = sapply(1:nrow(dataGrid), function(i) {
       para = as.list(dataGrid[i, -c(1, which(is.na(dataGrid[i,
           ]))), drop = FALSE])
       str = as.character(call(dataGrid[i, 1], para))
       str[2] = substr(str[2], start = 6, stop = nchar(str[2]))
       tmp = "("
       if (dataArgument) {
           if (nchar(str[2]) == 1) {
               tmp = "(d1"
           }
           else {
               tmp = "(d1, "
           }
       }
       str[2] = paste(tmp, str[2], sep = "")
       paste(str, collapse = "")
   })
   hc = paste(paste(var, 1:length(s), sep = ""), "<-", s)
   if (!silent)
       cat(hc, "\\n")
   hc
 }
}
\author{
who you are
}
\references{
~put references to the literature/web site here ~
}
\seealso{
objects to See Also as \code{\link{help}}, ~
}
\keyword{~kwd1}
\keyword{~kwd2}

