% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.eval_tibbles.R
\name{print.eval_tibbles}
\alias{print.eval_tibbles}
\title{Printing simulation results}
\usage{
\method{print}{eval_tibbles}(x)
}
\arguments{
\item{x}{object of class \code{eval_tibbles}}
}
\description{
Prints objects created by \code{eval_tibbles()}
}
\author{
Marsel Scheer
}
