% Generated by roxygen2 (4.0.1): do not edit by hand
\name{expandGrid}
\alias{expandGrid}
\title{Creates a data.frame from All Combinations}
\usage{
expandGrid(...)
}
\arguments{
\item{\dots}{vectors, factors or a list containing these.}
}
\value{
See \code{\link{expand.grid}}
}
\description{
Actually a wrapper for \code{\link{expand.grid}}, but
 character vectors will not be coerced to factors.
}
\examples{
expandGrid(fun="rnorm", mean=1:4, sd=2:5)
}
\author{
Marsel Scheer
}
\seealso{
\code{\link{expand.grid}}
}

