% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.eval_tibbles.R
\name{print.eval_tibbles}
\alias{print.eval_tibbles}
\title{Printing simulation results}
\usage{
\method{print}{eval_tibbles}(x, ...)
}
\arguments{
\item{x}{object of class \code{eval_tibbles}}

\item{...}{not used. only necessary to define the function consistently
with respect to \code{print(x, ...)}}
}
\description{
Prints objects created by \code{eval_tibbles()}
}
\author{
Marsel Scheer
}
\keyword{internal}
